requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
google-generativeai>=0.5.0
protobuf
pdfplumber==0.10.3
Pillow
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
email-validator==2.1.0.post1